split functions to multiple files in queries
- use require_once instead on include_once
